Extract date parsing to DateParserUtil
