+++ src/main/java/seedu/address/storage/JsonAdaptedEvent.java
@@ -10,7 +10,7 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.logic.parser.ParserUtil;
+import seedu.address.logic.parser.DateParserUtil;
 import seedu.address.model.event.Event;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.UniquePersonList;
@@ -88,7 +88,7 @@ class JsonAdaptedEvent {
             throw new IllegalValueException(String.format(MISSING_FIELD_MESSAGE_FORMAT, "Date"));
         }
 
-        LocalDateTime parsedDate = ParserUtil.parseDate(date);
+        LocalDateTime parsedDate = DateParserUtil.parseDate(date);
 
         final List<Tag> eventTags = new ArrayList<>();
         for (JsonAdaptedTag tag : tags) {
